Split custom-column specs
